fix collateral tasks not being import and therefore shceduling of them failing
