+++ validator/app/src/compute_horde_validator/validator/clean_me_up.py
@@ -1,5 +1,12 @@
 import asyncio
+import functools
+import time
+from collections.abc import Callable
+from typing import ParamSpec, TypeVar
 
+import turbobt
+from asgiref.sync import async_to_sync, sync_to_async
+from bt_ddos_shield.turbobt import ShieldedBittensor
 from celery.utils.log import get_task_logger
 from compute_horde.miner_client.organic import OrganicMinerClient
 from compute_horde.transport.base import TransportConnectionError
@@ -63,3 +70,55 @@ async def get_single_manifest(
         )
         logger.warning(msg)
         return hotkey, None
+
+
+async def _get_metagraph_for_sync(bittensor: turbobt.Bittensor, block_number=None):
+    try:
+        start_ts = time.time()
+        subnet = bittensor.subnet(settings.BITTENSOR_NETUID)
+
+        async with bittensor.block(block_number) as block:
+            neurons, subnet_state = await asyncio.gather(subnet.list_neurons(), subnet.get_state())
+
+        duration = time.time() - start_ts
+        msg = f"Metagraph fetched: {len(neurons)} neurons @ block {block.number} in {duration:.2f} seconds"
+        logger.info(msg)
+        await SystemEvent.objects.using(settings.DEFAULT_DB_ALIAS).acreate(
+            type=SystemEvent.EventType.METAGRAPH_SYNCING,
+            subtype=SystemEvent.EventSubType.SUCCESS,
+            long_description=msg,
+            data={"duration": duration},
+        )
+        return neurons, subnet_state, block
+    except Exception as e:
+        msg = f"Failed to fetch neurons: {e}"
+        logger.warning(msg)
+        await SystemEvent.objects.using(settings.DEFAULT_DB_ALIAS).acreate(
+            type=SystemEvent.EventType.METAGRAPH_SYNCING,
+            subtype=SystemEvent.EventSubType.SUBTENSOR_CONNECTIVITY_ERROR,
+            long_description=msg,
+            data={},
+        )
+    return None, None, None
+
+
+P = ParamSpec("P")
+R = TypeVar("R")
+
+
+def bittensor_client(func: Callable[P, R]) -> Callable[..., R]:
+    @async_to_sync
+    async def synced_bittensor(*args, bittensor=None, **kwargs):
+        async with ShieldedBittensor(
+            settings.BITTENSOR_NETWORK,
+            ddos_shield_netuid=settings.BITTENSOR_NETUID,
+            ddos_shield_options=settings.BITTENSOR_SHIELD_METAGRAPH_OPTIONS(),
+            wallet=settings.BITTENSOR_WALLET(),
+        ) as bittensor:
+            return await sync_to_async(func)(*args, bittensor=bittensor, **kwargs)
+
+    @functools.wraps(func)
+    def wrapper(*args, **kwargs):
+        return synced_bittensor(*args, **kwargs)
+
+    return wrapper
